+++ components-ui/app/src/tol-ui/src/constants/cell-renderers.constants.ts
@@ -1,0 +1,11 @@
+/*
+SPDX-FileCopyrightText: 2026 Genome Research Ltd.
+
+SPDX-License-Identifier: MIT
+*/
+
+export const CELL_RENDERER_PARAMS = {
+  STRING: "string",
+  CONDITION: "condition",
+  MARKDOWN: "markdown"
+} as const;