+++ components-ui/app/src/tol-ui/src/interfaces/Table.ts
@@ -4,10 +4,14 @@ SPDX-FileCopyrightText: 2025 Genome Research Ltd.
 SPDX-License-Identifier: MIT
 */
 
-import { FieldMeta } from "..";
+import type { IFieldMeta } from "..";
+
+export type ITableRecord = Record<string, any>;
+
+export type ITableData = ITableRecord[];
 
 export interface ITableDrawerSave {
-  fieldMeta?: FieldMeta;
+  fieldMeta?: IFieldMeta;
   actions?: string[];
   defaultSortByAttribute?: string;
   defaultSortByType?: string;