+++ components-ui/app/src/tol-ui/src/table/RemoteTable.tsx
@@ -10,7 +10,7 @@ import {
   ActionCheckModal,
   ActionModal,
   API_METHODS,
-  FieldMeta,
+  IFieldMeta,
   IRemoteTargetAndZone,
   PUtilityBar,
   IZone,
@@ -59,7 +59,7 @@ export interface PRemoteTable extends IRemoteTargetAndZone, IHeight {
   /**
    * Initial field metadata for columns; overridden by any saved configuration for this table `id`
    */
-  fields?: FieldMeta;
+  fields?: IFieldMeta;
   /**
    * Default sort attribute when no saved sort configuration exists
    */
@@ -245,7 +245,7 @@ export function RemoteTable(props: PRemoteTable) {
 
   // data and field information
   const [data, setData] = useState<any[]>([]);
-  const [fieldMeta, setFieldMeta] = useState<FieldMeta>(
+  const [fieldMeta, setFieldMeta] = useState<IFieldMeta>(
     initialiseFieldMeta(
       getTableConfigLocalStorage(id, "fieldMeta") || fields
     )