+++ components-ui/app/src/tol-ui/src/table/Table.tsx
@@ -13,7 +13,7 @@ import {
   UtilityBar,
   resizeListener,
   ColumnConfigDrawer,
-  FieldMeta,
+  IFieldMeta,
   IDropdownButtonConfig,
   useStateFallback,
   IRemoteTargetAndZone,
@@ -37,7 +37,7 @@ import {
 export interface PTable extends IRemoteTargetAndZone {
   id: string;
   data: any;
-  fieldMeta: FieldMeta;
+  fieldMeta: IFieldMeta;
   height: any;
   loading: boolean;
   resizeableColumns?: boolean;