+++ CHANGELOG.md
@@ -6,6 +6,19 @@ SPDX-License-Identifier: MIT
 
 # Changelog
 
+## tol-ui `3.5.8`
+
+16-06-2026
+
+- Added the Card cell renderer and Markdown parameter type
+
+## tol-ui `3.5.7`
+
+12-06-2026
+
+- Added the DOWNLOAD button type
+- Added the `cleanable` and `searchable` SingleSelect options
+
 ## tol-ui `3.5.6`
 
 28-05-2026